added surrogate_data_persistor that persists images of trees with corresponding fitnesses.
